arm64: defconfig: use flatmem instead of sparsemem
Currently we didn't use non-continuous memory, also memory hotplug was not used,
so just change to flatmem model.

+++ include/asm-generic/memory_model.h
@@ -28,7 +28,7 @@
 #if defined(CONFIG_FLATMEM)
 
 #define __pfn_to_page(pfn)	(mem_map + ((pfn) - ARCH_PFN_OFFSET))
-#define __page_to_pfn(page)	((unsigned long)((page) - mem_map) + \
+#define __page_to_pfn(page)	(unsigned long)((unsigned long)((page) - mem_map) + \
 				 ARCH_PFN_OFFSET)
 #elif defined(CONFIG_DISCONTIGMEM)
